FIX: move RED_MOCK to test before
Since the Red Mock needs to mock which nodes it will listen for and return those nodes, they can and will be different for each test.

Setting it before the test will overwrite each other and allow tests to influence one another.

Better to move this to the before call.