UPDATE: add in max listener count
The listener count is an implementation detail for connecting to salesforce. It is helpful to ensure we don’t have loops creating connections, but not to the spirit of getting connections to salesforce. Need to determine a better way to allow it to be configured, where it isn’t front and center in the configurations of the node…